New logger added
