+++ pkg/utils/utils_test.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"testing"
 
-	localstate "github.com/kubesimplify/ksctl/internal/storagelogger/local"
+	localstate "github.com/kubesimplify/ksctl/internal/storage/local"
 	"github.com/kubesimplify/ksctl/pkg/resources"
 
 	. "github.com/kubesimplify/ksctl/pkg/utils/consts"
